requirements.txt
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
polars>=0.18.0
matplotlib>=3.5.0
seaborn>=0.12.0
scipy>=1.10.0
